+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2026-07-14
+
 ### Changed
 - Align inbound C-STORE with pynetdicom `event.dataset` + `event.file_meta`: expose `StoreRequest.FileMeta`, remove convenience `SaveAs`/`File`